Rag-agent/requirements.txt
playwright
playwright-stealth
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
tqdm
fastapi
uvicorn
